Add a local VOD library demo and fix fMP4 parse/write.
Parse MEDIA-SEQUENCE as long and PDT fractions up to 9 digits so Unified
Channel 1 snapshots work. Re-announce EXT-X-MAP after DISCONTINUITY.

Demo: serve /media/, Library rail, posters, JSON decision logs, seek
thumbnails, and absolute /media/ URLs on session create. GIFF Day 1 is
the library house ad at 10s. Backlog SSAI bufferAppendError when the ad
container or init does not match the title.

+++ library/src/main/java/com/iheartradio/m3u8/MediaParseState.java
@@ -12,7 +12,7 @@ class MediaParseState implements PlaylistParseState<MediaPlaylist> {
     public final List<TrackData> tracks = new ArrayList<>();
 
     public Integer targetDuration;
-    public Integer mediaSequenceNumber;
+    public Long mediaSequenceNumber;
     public boolean isIframesOnly;
     public PlaylistType playlistType;
     public TrackInfo trackInfo;
